# app/schemas/charity_project.py
from datetime import datetime
from typing import Optional

from pydantic import BaseModel, ConfigDict, Field


class CharityProjectBase(BaseModel):
    """Базовая схема проекта."""

    name: str = Field(..., min_length=5, max_length=100)
    description: str = Field(..., min_length=10)
    full_amount: int = Field(..., gt=0)

    model_config = ConfigDict(extra='forbid')


class CharityProjectCreate(CharityProjectBase):
    """Схема для создания проекта."""

    pass


class CharityProjectUpdate(BaseModel):
    """Схема для обновления проекта."""

    name: Optional[str] = Field(None, min_length=5, max_length=100)
    description: Optional[str] = Field(None, min_length=10)
    full_amount: Optional[int] = Field(None, gt=0)

    model_config = ConfigDict(extra='forbid')


class CharityProjectDB(CharityProjectBase):
    """Схема для ответа."""

    id: int
    invested_amount: int = 0
    fully_invested: bool = False
    create_date: datetime
    close_date: Optional[datetime] = None

    model_config = ConfigDict(from_attributes=True)